// client/Magnus-Laser/src/graphql/types.ts
export type Maybe<T> = T | null | undefined;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};
// ...
export type Building = {
  __typename?: 'Building';
  ID: Scalars['ID']['output'];
  backupLights: Scalars['Boolean']['output'];
  description: Scalars['String']['output'];
  elevators: Scalars['Boolean']['output'];
  emergencyExit: Scalars['Boolean']['output'];
  event: BuildingEvent;
  gatehouseFrontDesk: Scalars['Boolean']['output'];
  image: Scalars['String']['output'];
  isAbandoned: Scalars['Boolean']['output'];
  landingPad: Scalars['Boolean']['output'];
  name: Scalars['String']['output'];
  ownership: BuildingOwnership;
  parking: Scalars['Boolean']['output'];
  secret: BuildingSecret;
  secretOrAltEntrance: Scalars['Boolean']['output'];
  securityPersonnel: BuildingSecurityPersonnel;
  style: BuildingStyle;
  type: BuildingType;
};

export type BuildingComplication = {
  __typename?: 'BuildingComplication';
  character?: Maybe<Character>;
  item?: Maybe<Item>;
  type: PlotBuildingComplicationType;
};

export enum BuildingEvent {
  CRUMBLING_DEMOLITION = 'CRUMBLING_DEMOLITION',
  EDGERUNNERS_GOING_SAME_PLACE = 'EDGERUNNERS_GOING_SAME_PLACE',
  ELITE_CREW_VISITING = 'ELITE_CREW_VISITING',
  GUN_FIGHT = 'GUN_FIGHT',
  MAINTENANCE_PROBLEM = 'MAINTENANCE_PROBLEM',
  NETRUNNER_MESSING_WITH_SYSTEMS = 'NETRUNNER_MESSING_WITH_SYSTEMS',
  ONGOING_KIDNAP = 'ONGOING_KIDNAP',
  OWNERS_COLLECTING_RENT = 'OWNERS_COLLECTING_RENT',
  PRIVATE_INVESTIGATOR = 'PRIVATE_INVESTIGATOR',
  STRIKE = 'STRIKE',
  STRONG_SECURITY_PRESENCE = 'STRONG_SECURITY_PRESENCE',
  WORKERS_REMODELING = 'WORKERS_REMODELING'
}

export enum BuildingOwnership {
  BYSTANDER = 'BYSTANDER',
  CORPO = 'CORPO',
  FIXER = 'FIXER',
  GANG_MAFIA = 'GANG_MAFIA',
  GOVERNMENT = 'GOVERNMENT',
  LOCAL_GOV = 'LOCAL_GOV',
  LOW_LEVEL_GONKS = 'LOW_LEVEL_GONKS',
  MEGA_CORPO = 'MEGA_CORPO',
  MILITARISTIC_GANG = 'MILITARISTIC_GANG',
  NO_ONE_SCAVS = 'NO_ONE_SCAVS',
  POSERGANG = 'POSERGANG',
  SMALL_BUSINESS = 'SMALL_BUSINESS'
}

export enum BuildingSecret {
  ACTIVIST_SABOTAGE = 'ACTIVIST_SABOTAGE',
  COVERT_OP_MEETING_POINT = 'COVERT_OP_MEETING_POINT',
  DATA_STORE_NETRUNNER_DEN = 'DATA_STORE_NETRUNNER_DEN',
  DRUG_LAB = 'DRUG_LAB',
  DRUG_STASH = 'DRUG_STASH',
  FIXER_ARRANGEMENT = 'FIXER_ARRANGEMENT',
  MEDIA_INVESTIGATION = 'MEDIA_INVESTIGATION',
  PARTY = 'PARTY',
  SAFE_HOUSE = 'SAFE_HOUSE',
  SECRET_SCIENCE_LAB = 'SECRET_SCIENCE_LAB',
  SOMEONE_KIDNAPPED = 'SOMEONE_KIDNAPPED',
  WEAPON_STASH = 'WEAPON_STASH'
}

export enum BuildingSecurityPersonnel {
  BORG = 'BORG',
  CITY_SEC = 'CITY_SEC',
  CORPO_SEC = 'CORPO_SEC',
  ELITE_TROOPS = 'ELITE_TROOPS',
  FAST_RESPONSE_BACKUP = 'FAST_RESPONSE_BACKUP',
  HEAVY_VEHICLES = 'HEAVY_VEHICLES',
  HEAVY_WEAPONS = 'HEAVY_WEAPONS',
  LOCALS_TENNANTS = 'LOCALS_TENNANTS',
  LOCAL_SEC_GANG = 'LOCAL_SEC_GANG',
  NONE = 'NONE',
  RESPONSE_BACKUP = 'RESPONSE_BACKUP',
  VEHICLES = 'VEHICLES'
}

export enum BuildingStyle {
  AUSTERE = 'AUSTERE',
  CORPORATE = 'CORPORATE',
  EUROPEAN = 'EUROPEAN',
  EXOTIC = 'EXOTIC',
  LUXURIOUS = 'LUXURIOUS',
  MILITARISTIC = 'MILITARISTIC',
  MODERN = 'MODERN',
  NEON_FEST = 'NEON_FEST',
  ORIENTAL = 'ORIENTAL',
  SOVIETIC = 'SOVIETIC',
  TRIBAL = 'TRIBAL',
  URBAN_GRAFFITI = 'URBAN_GRAFFITI'
}

export enum BuildingType {
  ABANDONED_BUILDING = 'ABANDONED_BUILDING',
  COMMERCIAL_BUILDING = 'COMMERCIAL_BUILDING',
  CORPO_BUILDING = 'CORPO_BUILDING',
  CUBE_HOTEL_MOTEL_CARGO_CONTAINER = 'CUBE_HOTEL_MOTEL_CARGO_CONTAINER',
  ESTABLISHMENT = 'ESTABLISHMENT',
  GOV_BUILDING = 'GOV_BUILDING',
  LUXURY_PENTHOUSE_MCMANSION = 'LUXURY_PENTHOUSE_MCMANSION',
  MEGACORPO_HQ = 'MEGACORPO_HQ',
  MULTI_STORY_BUILDING = 'MULTI_STORY_BUILDING',
  PUBLIC_SPACE = 'PUBLIC_SPACE',
  SKYSCRAPER_MEGABUILDING = 'SKYSCRAPER_MEGABUILDING',
  VACANT_LOT_CONSTRUCTION_SITE = 'VACANT_LOT_CONSTRUCTION_SITE'
}

export type Character = {
  __typename?: 'Character';
  ID: Scalars['ID']['output'];
  attitude: CharacterAttitude;
  image: Scalars['String']['output'];
  name: Scalars['String']['output'];
  type: CharacterType;
};

export enum CharacterAttitude {
  COLD_AND_PROFESSIONAL = 'COLD_AND_PROFESSIONAL',
  CONSIDERS_A_CREW_MEMBER_A_GOOD_FRIEND = 'CONSIDERS_A_CREW_MEMBER_A_GOOD_FRIEND',
  CONSIDERS_A_CREW_MEMBER_A_PARTNER = 'CONSIDERS_A_CREW_MEMBER_A_PARTNER',
  DOESN_T_TRUST_THE_CREW = 'DOESN_T_TRUST_THE_CREW',
  GENERALLY_DISLIKES_THE_CREW = 'GENERALLY_DISLIKES_THE_CREW',
  NEUTRAL_ATTITUDE_TOWARDS_THE_CREW = 'NEUTRAL_ATTITUDE_TOWARDS_THE_CREW',
  THINKS_THE_CREW_COULD_BE_POTENTIAL_ALLIES = 'THINKS_THE_CREW_COULD_BE_POTENTIAL_ALLIES',
  TRUSTS_IN_THE_CREW_S_DECISIONS = 'TRUSTS_IN_THE_CREW_S_DECISIONS',
  WANTS_A_CREW_MEMBER_DEAD = 'WANTS_A_CREW_MEMBER_DEAD',
  WORSHIPS_A_CREW_MEMBER = 'WORSHIPS_A_CREW_MEMBER'
}

export enum CharacterType {
  AVERAGE_CITIZEN_STREETRAT = 'AVERAGE_CITIZEN_STREETRAT',
  CELEBRITY = 'CELEBRITY',
  CORPORATE_EXEC = 'CORPORATE_EXEC',
  GOVERNMENT_OFFICIAL = 'GOVERNMENT_OFFICIAL',
  MERCENARY_FREELANCER_ASSASSIN = 'MERCENARY_FREELANCER_ASSASSIN',
  NETRUNNER_HACKER = 'NETRUNNER_HACKER',
  NOMAD = 'NOMAD',
  POLICE_LAWMAN = 'POLICE_LAWMAN',
  POLITICIAN = 'POLITICIAN',
  POP_STAR = 'POP_STAR',
  RELIGIOUS_FIGURE = 'RELIGIOUS_FIGURE',
  RIPPERDOC = 'RIPPERDOC',
  TECH = 'TECH',
  TERRORIST = 'TERRORIST'
}
// ...
export type FixerJob = {
  __typename?: 'FixerJob';
  ID: Scalars['ID']['output'];
  description: Scalars['String']['output'];
  difficulty: JobDifficulty;
  image: Scalars['String']['output'];
  name: Scalars['String']['output'];
  plot: Plot;
};
// ...
export type Item = {
  __typename?: 'Item';
  ID: Scalars['ID']['output'];
  condition: ItemCondition;
  image: Scalars['String']['output'];
  name: Scalars['String']['output'];
  type: ItemType;
};

export enum ItemCondition {
  AVERAGE_QUALITY = 'AVERAGE_QUALITY',
  BELOW_AVERAGE_QUALITY = 'BELOW_AVERAGE_QUALITY',
  BRAND_NEW_AND_IN_GREAT_CONDITION = 'BRAND_NEW_AND_IN_GREAT_CONDITION',
  BROKEN_AND_IN_PIECES = 'BROKEN_AND_IN_PIECES',
  COLLECTORS_ITEM_FEW_IN_EXISTENCE = 'COLLECTORS_ITEM_FEW_IN_EXISTENCE',
  ENHANCED_IN_SOME_WAY_AND_SOUGHT_AFTER = 'ENHANCED_IN_SOME_WAY_AND_SOUGHT_AFTER',
  FUNCTIONAL_AND_UTILITARIAN = 'FUNCTIONAL_AND_UTILITARIAN',
  GOOD_CONDITION_BUT_USED = 'GOOD_CONDITION_BUT_USED',
  LOW_GRADE_CHEAP_MATERIALS = 'LOW_GRADE_CHEAP_MATERIALS',
  ONE_OF_A_KIND_ITEM_TOP_QUALITY = 'ONE_OF_A_KIND_ITEM_TOP_QUALITY'
}

export enum ItemType {
  AI_ROBOT_DRONE = 'AI_ROBOT_DRONE',
  BIOLOGICAL_SAMPLES = 'BIOLOGICAL_SAMPLES',
  CYBERWARE = 'CYBERWARE',
  DIGITAL_FILES = 'DIGITAL_FILES',
  DRUGS_ILLEGAL_CONTRABAND = 'DRUGS_ILLEGAL_CONTRABAND',
  EXOTIC_ANIMAL = 'EXOTIC_ANIMAL',
  FOOD_FUELS_SUPPLIES = 'FOOD_FUELS_SUPPLIES',
  MONEY = 'MONEY',
  VEHICLE = 'VEHICLE',
  WEAPONS = 'WEAPONS'
}

export enum ItemVerb {
  DELIVER = 'DELIVER',
  DESTROY = 'DESTROY',
  HIDE = 'HIDE',
  MODIFY = 'MODIFY',
  STEAL = 'STEAL',
  TRADE = 'TRADE',
  USE = 'USE'
}

export type ItemVerbWrapper = Verb & {
  __typename?: 'ItemVerbWrapper';
  value: ItemVerb;
};

export enum JobDifficulty {
  DANGEROUS = 'DANGEROUS',
  EASY = 'EASY',
  TYPICAL = 'TYPICAL'
}
// ...
export type Plot = {
  __typename?: 'Plot';
  plotBuilding: PlotBuilding;
  plotComplication: PlotComplication;
  plotSubject?: Maybe<PlotSubject>;
  verb: PlotVerb;
};

export type PlotBuilding = {
  __typename?: 'PlotBuilding';
  building: Building;
  complication: BuildingComplication;
};

export enum PlotBuildingComplicationType {
  ACCIDENT_ZONE_INVOLVES_SHIPMENT_OF_ITEM = 'ACCIDENT_ZONE_INVOLVES_SHIPMENT_OF_ITEM',
  CCTV_POLICE_SURVEILLANCE_IS_HEAVY = 'CCTV_POLICE_SURVEILLANCE_IS_HEAVY',
  CONCERT_HOSTED_BY_CHARACTER = 'CONCERT_HOSTED_BY_CHARACTER',
  FORTIFIED_POSITIONS_BY_SECURITY_PERSONNEL_BECAUSE_OF_ITEM = 'FORTIFIED_POSITIONS_BY_SECURITY_PERSONNEL_BECAUSE_OF_ITEM',
  HEAVY_TRAFFIC = 'HEAVY_TRAFFIC',
  NEED_ID_PASSWORD_TO_GET_IN_HELD_BY_CHARACTER = 'NEED_ID_PASSWORD_TO_GET_IN_HELD_BY_CHARACTER',
  NO_GUN_ZONE = 'NO_GUN_ZONE',
  RECENTLY_TAGGED_AND_LOOTED = 'RECENTLY_TAGGED_AND_LOOTED',
  SHOOTOUT_WITH_POLICE_AND_CYBERPSYCHO_CHARACTER = 'SHOOTOUT_WITH_POLICE_AND_CYBERPSYCHO_CHARACTER',
  TRAUMA_TEAM_ON_THE_SCENE = 'TRAUMA_TEAM_ON_THE_SCENE'
}

export enum PlotBuildingVerb {
  DEFEND = 'DEFEND',
  INVESTIGATE = 'INVESTIGATE',
  LOOT = 'LOOT',
  OCCUPY = 'OCCUPY',
  VANDALIZE = 'VANDALIZE'
}

export type PlotBuildingVerbWrapper = Verb & {
  __typename?: 'PlotBuildingVerbWrapper';
  value: PlotBuildingVerb;
};

export type PlotComplication = {
  __typename?: 'PlotComplication';
  character?: Maybe<Character>;
  item?: Maybe<Item>;
  type: PlotComplicationType;
};
// ...
export type PlotSubject = Character | Item | PlotGang;

export type PlotVerb = CharacterVerbWrapper | ItemVerbWrapper | PlotBuildingVerbWrapper | PlotGangVerbWrapper;